requirements.txt
streamlit>=1.32.0
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.36.0
timm>=0.9.12
numpy>=1.24.0
opencv-python-headless>=4.8.0
matplotlib>=3.7.0
Pillow>=10.1.0
scikit-learn>=1.3.0
plotly>=5.18.0
scipy>=1.11.0
pandas>=2.1.0
pyttsx3>=2.90
